+++ android/app/build.gradle
@@ -88,7 +88,7 @@ android {
     buildToolsVersion "23.0.1"
 
     defaultConfig {
-        applicationId "com.brewbaseclient"
+        applicationId "com.worrywortclient"
         minSdkVersion 16
         targetSdkVersion 22
         versionCode 1
