+++ android/app/src/main/java/com/worrywortclient/MainActivity.java
@@ -1,4 +1,4 @@
-package com.brewbaseclient;
+package com.worrywortclient;
 
 import com.facebook.react.ReactActivity;
 
@@ -10,6 +10,6 @@ public class MainActivity extends ReactActivity {
      */
     @Override
     protected String getMainComponentName() {
-        return "BrewbaseClient";
+        return "WorryWortClient";
     }
 }
