+++ android/settings.gradle
@@ -1,4 +1,4 @@
-rootProject.name = 'BrewbaseClient'
+rootProject.name = 'WorryWortClient'
 include ':react-native-config'
 project(':react-native-config').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-config/android')
