+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "BrewbaseClient",
-  "displayName": "BrewbaseClient"
+  "name": "WorryWortClient",
+  "displayName": "WorryWortClient"
 }
