+++ ios/WorryWortClient/AppDelegate.h
