+++ ios/WorryWortClient/Images.xcassets/AppIcon.appiconset/Contents.json
