+++ ios/WorryWortClientTests/Info.plist
