+++ ios/WorryWortClientTests/WorryWortClientTests.m
@@ -16,11 +16,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React Native!"
 
-@interface BrewbaseClientTests : XCTestCase
+@interface WorryWortClientTests : XCTestCase
 
 @end
 
-@implementation BrewbaseClientTests
+@implementation WorryWortClientTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL(^)(UIView *view))test
 {
