+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "BrewbaseClient",
+  "name": "WorryWortClient",
   "version": "0.0.1",
   "private": true,
   "scripts": {
